requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-wtf
email-validator
flask-login
pydenticon
python-dotenv